Only sort first pragma block and allow support different header orders.
This change finds the (first) pragma block explicitly instead of
assuming that all pragmas appear at the top of the file.  This allows
things like the module header comments to come either before or after
the pragma sections.

This change also updates the pragma sorting to only sort within that
first pragma block.  Previously, sorting was applied to all lines from
the beginning of the file to the module statement which would, for
example, re-sort module comments along with the pragmas.